requirements.txt
timm==0.4.12
transformers==4.15.0
fairscale==0.4.4
pycocoevalcap
fastapi
uvicorn
python-dotenv
python-multipart
